# docs/swagger.yaml
basePath: /api
definitions:
  dashboard.DashboardResponse:
    properties:
      balance:
        type: number
      total_expense:
        type: number
      total_income:
        type: number
    type: object
  family.CreateFamilyRequest:
    properties:
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    type: object
  family.FamilyMemberResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  family.GetMyFamilyResponse:
    properties:
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  family.InvitationResponse:
    properties:
      family_id:
        type: string
      family_name:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
  family.InviteMemberRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.Pagination:
    properties:
      has_more:
        type: boolean
      next_cursor:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorDetail'
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      success:
        type: boolean
    type: object
  transaction.CreateTransactionRequest:
    properties:
      amount:
        type: number
      description:
        type: string
      family_id:
        type: string
      title:
        type: string
      type:
        enum:
        - income
        - expense
        type: string
    required:
    - family_id
    - title
    - type
    type: object
  transaction.CreateTransactionResponse:
    properties:
      id:
        type: string
    type: object
  transaction.GetTransactionsResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      type:
        type: string
      user_name:
        type: string
    type: object
info:
  contact: {}
  description: Family Expense Tracker API
  title: Evara API
  version: "1.0"
paths:
  /dashboard:
    get:
      description: Get dashboard summary for a family
      parameters:
      - description: Family ID
        in: query
        name: family_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dashboard.DashboardResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get dashboard
      tags:
      - Dashboard
  /families:
    post:
      consumes:
      - application/json
      description: Create a new family
      parameters:
      - description: Create Family
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/family.CreateFamilyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create family
      tags:
      - Family
  /families/{id}/invite:
    post:
      consumes:
      - application/json
      description: Invite a user to join a family by email
      parameters:
      - description: Family ID
        in: path
        name: id
        required: true
        type: string
      - description: Invitation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/family.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Invite family member
      tags:
      - Family
  /families/{id}/members:
    get:
      description: Get all members of a family
      parameters:
      - description: Family ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/family.FamilyMemberResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get family members
      tags:
      - Family
  /families/invitations:
    get:
      description: Get all pending family invitations for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/family.InvitationResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get invitations
      tags:
      - Family
  /families/invitations/{id}/accept:
    patch:
      description: Accept a pending family invitation
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Accept invitation
      tags:
      - Family
  /families/invitations/{id}/reject:
    patch:
      description: Reject a pending family invitation
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Reject invitation
      tags:
      - Family
  /families/me:
    get:
      description: Get the authenticated user's family information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/family.GetMyFamilyResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get my family
      tags:
      - Family
  /me:
    get:
      description: Get the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: objec
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - User
  /transaction:
    get:
      description: Get all transactions form a family
      parameters:
      - description: Family ID
        in: query
        name: family_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/transaction.GetTransactionsResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get transactions
      tags:
      - Transaction
    post:
      consumes:
      - application/json
      description: Create a new income or expense transaction
      parameters:
      - description: Transaction data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transaction.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/transaction.CreateTransactionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - Transaction
  /transaction/{id}:
    delete:
      description: Delete a transaction by its ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - Transaction
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"